Created the calendars page
For now i've created the function that generates and retrieves the proper dates for each month outside of this program in a scratch pad and copy pasted the data into months_days.py.  I will eventually move the function into the program so it is ran on page load, along with other functions that generate the following years dates and the previous years dates.